# docs/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Denguekay: pronóstico distrital de dengue",
    "description": "API del documento OE2 (Tabla 2) sobre datos agregados de Piura. Las predicciones de versiones que no cumplen los umbrales se marcan experimental: true (demo; desviación temporal documentada).",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/modelos/activo": {
      "get": {
        "tags": [
          "Modelos"
        ],
        "summary": "Versiones en uso por horizonte (HU0008)",
        "description": "Versión activa por tarea; sin activa, la selección experimental marcada ``experimental: true``.",
        "operationId": "modelos_activos_api_v1_modelos_activo_get",
        "parameters": [
          {
            "name": "horizonte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4,
                  "minimum": 2,
                  "description": "Anticipación en semanas: 2 o 4 con modelo; 3 responde no_disponible"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizonte"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelosActivosRespuesta"
                },
                "examples": {
                  "bd_real": {
                    "summary": "Ejemplo obtenido de la base de datos configurada",
                    "value": {
                      "disponible": true,
                      "motivo": null,
                      "fecha_corte_datos": "2025-12-27",
                      "fecha_actualizacion": "2026-10-02T04:59:55.927919Z",
                      "elementos": [
                        {
                          "id_version": 1,
                          "codigo": "clf-h2-v1",
                          "tarea": "clasificacion",
                          "horizonte": 2,
                          "algoritmo": "xgboost",
                          "estado": "candidata",
                          "cumple_umbrales": false,
                          "experimental": true,
                          "umbral_probabilidad": 0.2750000000000001,
                          "mlflow_run_id": "a3cc943abf4241bb82774d13b2792853",
                          "fecha_registro": "2026-10-02T04:59:36.364915Z",
                          "fecha_activacion": null,
                          "seleccion": "experimental"
                        },
                        {
                          "id_version": 2,
                          "codigo": "reg-h2-v1",
                          "tarea": "regresion",
                          "horizonte": 2,
                          "algoritmo": "xgboost",
                          "estado": "candidata",
                          "cumple_umbrales": false,
                          "experimental": true,
                          "umbral_probabilidad": null,
                          "mlflow_run_id": "a3cc943abf4241bb82774d13b2792853",
                          "fecha_registro": "2026-10-02T04:59:36.369988Z",
                          "fecha_activacion": null,
                          "seleccion": "experimental"
                        },
                        {
                          "id_version": 3,
                          "codigo": "clf-h4-v1",
                          "tarea": "clasificacion",
                          "horizonte": 4,
                          "algoritmo": "xgboost",
                          "estado": "candidata",
                          "cumple_umbrales": false,
                          "experimental": true,
                          "umbral_probabilidad": 0.22500000000000003,
                          "mlflow_run_id": "a3cc943abf4241bb82774d13b2792853",
                          "fecha_registro": "2026-10-02T04:59:36.372499Z",
                          "fecha_activacion": null,
                          "seleccion": "experimental"
                        },
                        {
                          "id_version": 4,
                          "codigo": "reg-h4-v1",
                          "tarea": "regresion",
                          "horizonte": 4,
                          "algoritmo": "xgboost",
                          "estado": "candidata",
                          "cumple_umbrales": false,
                          "experimental": true,
                          "umbral_probabilidad": null,
                          "mlflow_run_id": "a3cc943abf4241bb82774d13b2792853",
                          "fecha_registro": "2026-10-02T04:59:36.375840Z",
                          "fecha_activacion": null,
                          "seleccion": "experimental"
                        }
                      ],
                      "nota": "Demo experimental: ningún modelo cumple todavía los umbrales de aceptación; las predicciones marcadas experimental: true no son un modelo validado."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clave de escritura ausente o inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Solicitud no atendible con los datos o modelos (h=3, artefacto, umbrales)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Parámetro inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "503": {
            "description": "Base de datos o escritura no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Error interno controlado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/modelos/{identificador}/variables": {
      "get": {
        "tags": [
          "Modelos"
        ],
        "summary": "Importancias gain de una versión (extensión, HU0008-4)",
        "description": "Las importancias no implican causalidad; su ausencia se informa sin rellenar con ceros.",
        "operationId": "variables_api_v1_modelos__identificador__variables_get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Identificador positivo del registro guardado en la BD",
              "title": "Identificador"
            },
            "description": "Identificador positivo del registro guardado en la BD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariablesRespuesta"
                },
                "examples": {
                  "bd_real": {
                    "summary": "Ejemplo obtenido de la base de datos configurada",
                    "value": {
                      "disponible": true,
                      "motivo": null,
                      "fecha_corte_datos": "2025-12-27",
                      "fecha_actualizacion": "2026-10-02T04:59:55.927919Z",
                      "id_version": 1,
                      "elementos": [
                        {
                          "variable": "casos_media_4_h2",
                          "importancia": 221.74436950683594,
                          "rango": 1
                        },
                        {
                          "variable": "casos_lag_2",
                          "importancia": 173.24386596679688,
                          "rango": 2
                        },
                        {
                          "variable": "casos_semanas_positivas_4_h2",
                          "importancia": 53.17316818237305,
                          "rango": 3
                        },
                        {
                          "variable": "log_poblacion_censo_2017",
                          "importancia": 21.276065826416016,
                          "rango": 4
                        },
                        {
                          "variable": "semana_epi_coseno",
                          "importancia": 19.57233428955078,
                          "rango": 5
                        },
                        {
                          "variable": "casos_lag_3",
                          "importancia": 19.42538833618164,
                          "rango": 6
                        },
                        {
                          "variable": "semana_epi_seno",
                          "importancia": 10.286520004272461,
                          "rango": 7
                        }
                      ],
                      "nota": "Importancia gain del booster; no implica causalidad"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clave de escritura ausente o inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Solicitud no atendible con los datos o modelos (h=3, artefacto, umbrales)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Parámetro inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "503": {
            "description": "Base de datos o escritura no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Error interno controlado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorRespuesta": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "detalle": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalle"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "codigo",
          "mensaje"
        ],
        "title": "ErrorRespuesta",
        "description": "Error homogéneo, sin credenciales ni SQL interno."
      },
...
      "ModeloEnUso": {
        "properties": {
          "id_version": {
            "type": "integer",
            "title": "Id Version"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "tarea": {
            "type": "string",
            "enum": [
              "clasificacion",
              "regresion"
            ],
            "title": "Tarea"
          },
          "horizonte": {
            "type": "integer",
            "title": "Horizonte"
          },
          "algoritmo": {
            "type": "string",
            "title": "Algoritmo"
          },
          "estado": {
            "type": "string",
            "enum": [
              "candidata",
              "activa",
              "archivada",
              "rechazada"
            ],
            "title": "Estado"
          },
          "cumple_umbrales": {
            "type": "boolean",
            "title": "Cumple Umbrales"
          },
          "experimental": {
            "type": "boolean",
            "title": "Experimental"
          },
          "umbral_probabilidad": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Umbral Probabilidad"
          },
          "mlflow_run_id": {
            "type": "string",
            "title": "Mlflow Run Id"
          },
          "fecha_registro": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Registro"
          },
          "fecha_activacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Activacion"
          },
          "seleccion": {
            "type": "string",
            "enum": [
              "activa",
              "experimental"
            ],
            "title": "Seleccion",
            "description": "activa = estado 'activa'; experimental = parametro_sistema.seleccion_experimental"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id_version",
          "codigo",
          "tarea",
          "horizonte",
          "algoritmo",
          "estado",
          "cumple_umbrales",
          "experimental",
          "umbral_probabilidad",
          "mlflow_run_id",
          "fecha_registro",
          "fecha_activacion",
          "seleccion"
        ],
        "title": "ModeloEnUso",
        "description": "Versión en uso (activa o selección experimental) por tarea y horizonte"
      },
...
      "ModelosActivosRespuesta": {
        "properties": {
          "disponible": {
            "type": "boolean",
            "title": "Disponible",
            "description": "Indica que existe un valor utilizable; cero sí es un valor"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo",
            "description": "Explicación de la ausencia o restricción"
          },
          "fecha_corte_datos": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Corte Datos",
            "description": "Último cierre semanal de las fuentes cargadas"
          },
          "fecha_actualizacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Actualizacion",
            "description": "Fin de la última ejecución exitosa"
          },
          "elementos": {
            "items": {
              "$ref": "#/components/schemas/ModeloEnUso"
            },
            "type": "array",
            "title": "Elementos"
          },
          "nota": {
            "type": "string",
            "title": "Nota"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "disponible",
          "fecha_corte_datos",
          "fecha_actualizacion",
          "elementos",
          "nota"
        ],
        "title": "ModelosActivosRespuesta"
      },
...
      "VariablesRespuesta": {
        "properties": {
          "disponible": {
            "type": "boolean",
            "title": "Disponible",
            "description": "Indica que existe un valor utilizable; cero sí es un valor"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo",
            "description": "Explicación de la ausencia o restricción"
          },
          "fecha_corte_datos": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Corte Datos",
            "description": "Último cierre semanal de las fuentes cargadas"
          },
          "fecha_actualizacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Actualizacion",
            "description": "Fin de la última ejecución exitosa"
          },
          "id_version": {
            "type": "integer",
            "title": "Id Version"
          },
          "elementos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Elementos"
          },
          "nota": {
            "type": "string",
            "title": "Nota"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "disponible",
          "fecha_corte_datos",
          "fecha_actualizacion",
          "id_version",
          "elementos",
          "nota"
        ],
        "title": "VariablesRespuesta",
        "description": "Importancias gain reales; no constituyen relaciones causales"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Clave de escritura definida en API_KEY; las lecturas son públicas",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}